refactor: move SignUpForm and SignUpPrompt
