+++ src/components/GenericPagesComponents/LandingPageSignUpForm/index.tsx
@@ -1,0 +1,1 @@
+export { LandingPageSignUpForm } from "./LandingPageSignUpForm";
