+++ src/components/GenericPagesComponents/LandingPageSignupPrompt/index.tsx
@@ -1,0 +1,1 @@
+export { LandingPageSignupPrompt } from "./LandingPageSignupPrompt";
